Products
I created an product page. The page includes 3 inputs which are Name, Price and Description. Moreover, I put a button which called "Add" to add an product. When a product is added the inputs of the product and 2 buttons which are delete and edit comes together. In additon, while editting  is going on, add button turns to update button.