ProductDeleting
I created an product page. The page includes 3 inputs which are Name, Price and Description. Moreover, I put a button which called "Add" to add an product. When a product is added the inputs of the product and a button which is called delete comes together. If delete button is clicked, the product next to that delete button is deleting.